+++ InstanceData.java
@@ -3,20 +3,20 @@ import java.util.ArrayList;
 
 public class InstanceData {
 	public String instanceName;
-	public float fuelPrice;
-	public float fuelConsumptionEmptyTruckPerKm;
-	public float fuelConsumptionPerTonKm;
+	public double fuelPrice;
+	public double fuelConsumptionEmptyTruckPerKm;
+	public double fuelConsumptionPerTonKm;
 	public int laborCostperHour;
-	public float otherDistanceDependentCostsPerKm;
+	public double otherDistanceDependentCostsPerKm;
 	public int otherTimeDependentCostsPerKm;
-	public float timeTonService;
+	public double timeTonService;
 	public int revenue;
 	public ArrayList<Integer> nodesDepot;
 	public ArrayList<Integer> nodes;
 	public int volumeCap;
 	public int weightCap;
-	public float[][] times;
-	public float[][] distances;
+	public double[][] times;
+	public double[][] distances;
 	public int numberOfCities;
 	public int numberOfVehicles; 
 
@@ -25,14 +25,14 @@ public class InstanceData {
 		this.instanceName = datafile;
 	}
 	
-	public float getDistance (Node i, Node j){
+	public double getDistance (Node i, Node j){
 	int iLocation = i.location;
 	int jLocation = j.location;
 	return distances[iLocation-1][jLocation-1];
 	//System.out.println(distance);
 	}
 	
-	public float getTime (Node i, Node j){
+	public double getTime (Node i, Node j){
 	int iTime = i.location;
 	int jTime = j.location;
 //	System.out.println(iTime);
